build: poetry lock
Use the pin-reqs script to write two different lock files for now:
- the poetry-compatible one
- a native requirements.txt (so that we can install with other tools,
  like pixi or uv, while we transition out of poetry).

Additionally, I moved snakemake pinning to a constraints.txt file.
as we discussed in
https://github.com/omnibenchmark/omnibenchmark/pull/80
we'd like to keep an internal, PEP621-compatible list of dependencies
that is minimally pinned, and then test with various versions of some of
the dependencies.

This opens the way to, e.g., run the test suite with different versions
of snakemake.