chore: let the import name be the package name
this is an uncomfortable change to make (sorry!), but the current usage
of poetry was a sort of corner case that was not going to float when
moving to a regular install.

In a gist, we were installing the 'omnibenchmark' package, but upon
installation the import name was 'omni' (the included source). this was
causing inconveniences when running tests, for instance, since the
internal module hierarchy was full of references to a pacakge that was
not available for import.

I will follow-up with another comming in which I do the needed renames.