style: lint
I'd adding temporary fmt: off/on pragmas until we completely move to
ruff for linting.

Without this, black and ruff were fighting each other; the line breaking
of the long def lines was avoiding ruff to properly capture the noqa
pragmas.

We should move to ruff and stop using black, but I would rather do that
in a separate commit.