feat: pass timeout as resource annotation. Resolves #65 (#79)
This is passed verbatim to the subprocess invocation.

We use humanfriendly for passing timeout quantities in a flexible
format.

Initially I was using a custom "timeout" parameter in the resources
dict, but after discussion with @imallona during review of #79, we decided to support the
"runtime" key in the resources to pave the way for cluster executors.

Do note that, as stated in the code, we'll need to harmonize quantities
as minutes when passing "runtime" to slurm clusters (at least).