docs: add automated docs via `mkdocs` (#102)
This commit marks the transition of the live documentation site from the legacy custom build in renku gitlab to an integrated documentation site.

The idea is to facilitate public documentation living in sync with whatever features are merged into dev or main.

* Automate docs via mkdocs

* Update trigger to workflow dispatch

It remains pending to:

- Update domain to github pages.
- Settle on action triggers and update policy.
- Figure out how to do proper versioning of the docs.

commit squashed and merged by traven.