feat: pass --out-dir (Closes: #106)
Add a dedicated 'out_dir' flag to the CLI.

This feature is less clean than we'd like because of the following
reasons:
- we lack a good way of propagating variables to the context execution
  in the snakemake rules.
- we are invoiking the Benchmark constructor in different places. Not
  all of those need to be concerned with the actual execution
  environment (where the output location logically belongs).
- partially related: the current Benchmark object is conflating a "pure"
  representation of the benchmark data structure with unrelated
  contextual information about the _execution_ of the benchmark. We
  might take the opportunity when superseding the omni-schema model to
  have a better design around this.

With these caveats in mind, this commit allows use to specify different
output folders with minimal changes, but it should be noted the
difficulties in reasoning about the state associated with the Benchmark
object at different points of the execution.

Overall, this addition of the --out-dir flag is a practical improvement, but it
highlights deeper design challenges. Addressing these challenges will
require:

1. A clean mechanism for propagating variables to Snakemake.
2. Centralizing the creation of Benchmark objects to handle
execution-specific concerns.
3. Redesigning the Benchmark class to separate data representation from
execution context.

- Resolves: #106