bug: do not use runtime for the custom local timeout (#125)
Using 'runtime' under resources affects the knapsack calculation in
snakemake. For some reason, this also has the side-effect of disabling
the ability to specify a number of cores for the task execution, so we
were basically disabling parallelism.

The solution is to use a custom label passed in the resources.

Credit goes to @imallona for noticing that during testing of current dev branch.