feat: add extra process profiler, closes: #150
We add an extra profiler to compare with the data generated by the
snakemake benchmarking tool.

This profiler is only enabled if the dependency is available in the
environment, and it writes performance info in json alongside
bench_performance.txt. The purpose is that we can compare both and
validate.

The profiler we depend on will provide aggregation of load, mem, disk i/o
over a tree of child processes. It should not significantly impact
process performance, but it's good to keep an eye on it.

We include the profiler name and version to the perf summary.