feat: add general config file (Closes: #31) (#148)
* feat: add general config file (Closes: #31)

- Add general configuration accessor
- Document configuration file location and format
- Replace hardcoded paths with config-based lookups

* ci(tests): disable envmodules tests, takes more than 5min

I have spent enough time today trying to fix a test that randomly
started to misbehave.

For the record, and to resume work in the future, I changed to a more
contained strategy, i.e. use temporal folders for envmodules integration test.

I believe this is not a flaky test per se, but a poor choice of what to
test to ensure that envmodules are working. compiling binutils for more
than 5 min makes me think that perhaps compiling a simpler tool might be
a good enough assurance.

also, I'm starting to suspect that maybe we had cached intermediate
files (either in the runners cache or commited files to the repo?). I
didn't investigate further, but perhaps this is a possibility to explain
why this test was not failing before.

in any case, this is unrelated work for the purpose of the current PR.