feat: enable --task-timeout (#132)
* fix(feat): re-enable --task-timeout

we had temporarily disabled this to debug properly how it affects the
execution of a complex benchmark.

we need to be sure not to have a regression of the parallelism being
disabled by us adding custom resources that affect the DAG calculation.

* fix(docs): fix path to refs.bib
* Remove mamba and replace with conda miniforge (#152)

this solves the following error in osx tests in the CI:

error    libmamba Could not solve for environment specs
    The following packages are incompatible
    ├─ conda >=25.5 * is not installable because it requires
    │  ├─ conda-libmamba-solver >=24.11.0 *, which requires


Co-authored-by: Daniel Incicau <daniel.incicau@gmail.com>