feat: ob create benchmark & module (Closes: #171)
Add basica capabilities to create new benchmark and modules, using
pre-defined templates.

Adding copier as the main engine for templating - uses jinja templates.

In the future we can dettach the module logic from omnibenchmark
embedded (or online) repositories, and allow for specific logic defined
by each benchmark.

For now a lot of this is stub. However, it poses an interesting design
problem: how much information needs to be transferred from the stage
level to the module.

One important thing to do will be to template example tests - as I try
to do in this commit for demonstration. We could take these tests from
the validation specification - to be landed in the 0.4.x release cycle.
This would give us nice post-execution validation paths, and possibly
also the possibility to "test-drive" a module with, e.g.,
datasets+preprocessing that are taged as "example", or even to allow
automation in the sketched "ob try module ..." - since we can codegen
the population of the QA tests from the information contained in the
benchmark.yaml.

That said, a lesson learned from doing this PR is that we might be
lacking common structural information about the "contract" that all
modules in a given stage need to satisfy. Right now, the passing of
parameters is done on a module level - there's no interface to enforce
that is common to all modules in the same stage. If we had, we
could add part of this discovery mechanism to the templating logic
itself. To be discussed.

fixes

conditional template