ci: disable patch status enforcement
By request on the slack dev channel, I'm disabling the enforcement of
patch status for the time being.

We agreed that at the current dev point this is more of a nuisance,
since if forces us to write tests for other parts of the codebase we're
not directly modifying when writing a fix.

However, it's up to the PR reviewers to ask for a significantly higher
patch coverage on new features or code in the critical path.

We will review this policy after a few releases to see if it's worthy
to enforce patch coverage for new code again.

As such, reviewers are welcome to consider the inputs of the codecov
patch status report when reviewing code.