tests: fix mock_benchmark.yaml
Update the reference benchmark.

The reference benchmark needs to be updated to eliminate ambiguity in the outputs; and get the latest module commit.

Daniel ported the fix to all the open PRs in the queue. There's a separate PR for catching the error at the moment of parsing the YAML.

--
Co-authored by:

ben (wrote this commit message over daniel's blank one)