ci: run clustbench for every commit (Closes: #213)
As requested by @imallona, since we don't seem to fully trust the quality of the existing test suite.
An argument in favor of that is that recent changes to dev (de-omni-schema-fication of dev)
were "green" on unit tests but they trivially broke parsing of our flagship
"clustbench" example, which we all agree should not happen looking forward.

(Apparently I misunderstood the request to implement e2e tests recently on dev)

Let's just be aware that we're effectively coupling development to support the
running of whatever is found on the main branch of clustbench, plus the cascading side-effects out of our control
(like third party dependencies, build-ability of the provided environment recipes, service downtime on registries, etc).

Enforcing this policy should not be too hard since we're approaching a first public version with which we're offering some
guarantees of stability, at least concerning cli methods and YAML API. In that regard, we're possibly
doing well by sacrificing a big of development agility if that means catching potentially backward-incompatible
changes that would importunate our users.

That said, we probably should keep an eye on the execution time, and possibly
configure a custom runner for the repeated executions of clustbench on every PR.